move separate meshes operation into postoperations file remove separate meshes remove smr dropdown and exportmeshes button remove older toggleik and split objects scripts (no longer needed) move animation and map library files up a directory add note in manifest that the network is used for installing pillow library
